reafactor: move FromRequestParts impl of User to user.rs
